insert a nop instruction and then force the proofs through
the purpose here is to see how far these changes might propogate --
not very far really. I have admitted assertions in the relative
places, tagged with ASW in the comments so they can be located. There
are only 4 or 5 places where it's a problem but essentially we need to
be able to replace any instance of the modified code with the
original, I think, by way of some lemma and proceed from there.