mess around with decision procedure for register set equality.
I've been mostly just poking around, but this is the stat of something
we really need -- to be able to decide equality of machine
states. THat means registers and memory. And we have to be clever
about the PC since that will change -- it will have to point to the
same instruction, but the same number. not sure how that works.