make Peephole.v compile
through a mix of arbitrarily omitting instructions, and a little
mixing up of option types, get the symbolic execution stuff to
compile. There are quite a lot of instructions that are currently
unsupported pending a better model of locations. Also, I admitted a
default initial locs to get things to compile, but I think TomMD is
working on a solution to this bit...