a longer version of SymExpr_dec that doesn't work
I think I'm going about this the wrong way. This longer version
produces 670-odd goals of which a few are not provable. The other
version took forever to run, and tons of memory, but was provable. I
think we need another approach to this problem. Doesn't functional
extensionality give us more of this equality already?