modify locations to store more of a "trace" style structure.
This was simply a modification from a replacing a-list to a cons-ing
a-list. That is, an update to an already written value will cons the
new value onto the a-list instead of overwriting it. This may give the
behavior we want. Still need to factor out the registers into their
own store.